feat: Auth / Show error message #42
