+++ client/src/screens/user/Auth.module.css
@@ -23,9 +23,15 @@
     margin-bottom: 0.5rem;
 }
 
+.page .errors {
+    margin-top: 1.2rem;
+    font-size: 0.9rem;
+    color: var(--error-text);
+}
+
 .page .form {
     width: 100%;
-    margin-top: 3rem;
+    margin-top: 2rem;
 }
 
 .page .form input {
@@ -99,9 +105,4 @@ button.link {
     .form {
         width: 20rem;
     }
-    /* .bottom {
-        width: 100%;
-        height: 10vh;
-        background-color: aliceblue;
-    } */
 }
